backend/apps/teachers/serializers.py
from rest_framework import serializers

from apps.teachers.models import Lecture


class LectureSerializer(serializers.ModelSerializer):
    class Meta:
        model = Lecture
        fields = '__all__'