backend/requirements.txt
psycopg2-binary
psycopg
django
djangorestframework
django-cors-headers
gunicorn
requests
dotenv
celery==5.2.7
redis==4.4.0
pyyaml
pytz